# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JAC Hub API - Panel de gestión para estudio indie de videojuegos
servers:
  - url: /api
    description: Ruta base de la API
...
paths:
...
  /tareas:
    get:
      operationId: listTareas
      tags: [tareas]
      summary: Listar tareas
      parameters:
        - name: proyectoId
          in: query
          schema:
            type: integer
        - name: columna
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de tareas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tarea"
    post:
      operationId: createTarea
      tags: [tareas]
      summary: Crear tarea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TareaInput"
      responses:
        "201":
          description: Tarea creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tarea"
  /tareas/{id}:
    patch:
      operationId: updateTarea
      tags: [tareas]
      summary: Actualizar tarea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TareaUpdate"
      responses:
        "200":
          description: Tarea actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tarea"
    delete:
      operationId: deleteTarea
      tags: [tareas]
      summary: Eliminar tarea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Tarea eliminada
...
components:
  schemas:
...
    Tarea:
      type: object
      required: [id, titulo, columna, prioridad, fechaCreacion]
      properties:
        id:
          type: integer
        titulo:
          type: string
        descripcion:
          type: ["string", "null"]
        columna:
          type: string
        prioridad:
          type: string
        proyectoId:
          type: ["integer", "null"]
        asignadoA:
          type: ["string", "null"]
        etiquetas:
          type: array
          items:
            type: string
        fechaCreacion:
          type: string
        fechaLimite:
          type: ["string", "null"]
    TareaInput:
      type: object
      required: [titulo, columna]
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        columna:
          type: string
        prioridad:
          type: string
        proyectoId:
          type: integer
        asignadoA:
          type: string
        etiquetas:
          type: array
          items:
            type: string
        fechaLimite:
          type: string
    TareaUpdate:
      type: object
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        columna:
          type: string
        prioridad:
          type: string
        asignadoA:
          type: string
        etiquetas:
          type: array
          items:
            type: string
        fechaLimite:
          type: string
...