# lib/api-zod/src/generated/types/tarea.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAC Hub API - Panel de gestión para estudio indie de videojuegos
 * OpenAPI spec version: 0.1.0
 */

export interface Tarea {
  id: number;
  titulo: string;
  /** @nullable */
  descripcion?: string | null;
  columna: string;
  prioridad: string;
  /** @nullable */
  proyectoId?: number | null;
  /** @nullable */
  asignadoA?: string | null;
  etiquetas?: string[];
  fechaCreacion: string;
  /** @nullable */
  fechaLimite?: string | null;
}