# lib/api-zod/src/generated/types/tareaInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAC Hub API - Panel de gestión para estudio indie de videojuegos
 * OpenAPI spec version: 0.1.0
 */

export interface TareaInput {
  titulo: string;
  descripcion?: string;
  columna: string;
  prioridad?: string;
  proyectoId?: number;
  asignadoA?: string;
  etiquetas?: string[];
  fechaLimite?: string;
}